requirements.txt
alembic==1.7.5
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.5
MarkupSafe==2.0.1
pycparser==2.21
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
zipp==3.6.0
alembic==1.7.5
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.5
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
zipp==3.6.0
alembic==1.7.5
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.5
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
zipp==3.6.0
alembic==1.7.5
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
Mako==1.1.5
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
psycopg2-binary==2.9.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.27
threadpoolctl==3.0.0
Werkzeug==2.0.2
zipp==3.6.0
